Change naming for Collection metadata
