+++ evaluation/exploration.py
@@ -68,7 +68,7 @@ class Exploration:
                     ])
                     d2_table = copy.copy(d1_table)
                     d2_table.schema, d2_table.name, d2_table.rowcount = "d2_" + filename, "d2_" + filename, d1_table.rowcount - 1
-                    d1_metadata, d2_metadata = Collection([d1_table], "csv"), Collection([d2_table], "csv")
+                    d1_metadata, d2_metadata = CollectionMetadata([d1_table], "csv"), CollectionMetadata([d2_table], "csv")
 
                     self.neighbor_pair[filename] = [d1, d2, d1_metadata, d2_metadata]
                     self.visited.append(filename)
