+++ sdk/burdock/data/local_csv_adapter.py
@@ -1,6 +1,6 @@
 import pandas as pd
 
-from burdock.query.sql import MetadataLoader
+from burdock.metadata.collection import CollectionMetadata
 from burdock.query.sql.reader import PandasReader
 
 from .dataset_adapter import DatasetAdapter
@@ -18,7 +18,7 @@ class LocalCSVAdapter(DatasetAdapter):
         return pd.read_csv(dataset_document.csv_details.local_path)
 
     def _load_metadata(dataset_document):
-        return MetadataLoader(dataset_document.csv_details.local_path.split(".")[0] + ".yaml").read_schema()
+        return CollectionMetadata.from_file(dataset_document.csv_details.local_path.split(".")[0] + ".yaml")
 
     def _load_reader(dataset_document):
         return PandasReader(LocalCSVAdapter.load_metadata(dataset_document),
