+++ sdk/burdock/metadata/collection.py
@@ -1,9 +1,10 @@
 import importlib
+import yaml
 from .name_compare import BaseNameCompare
 
 # implements spec at https://docs.google.com/document/d/1Q4lUKyEu2W9qQKq6A0dbo0dohgSUxitbdGhX97sUNOM/
 
-class Collection:
+class CollectionMetadata:
     def __init__(self, tables, engine, compare=None):
         self.m_tables = dict([(t.table_name(), t) for t in tables])
         self.engine = engine if engine is not None else "Unknown"
@@ -25,6 +26,19 @@ class Collection:
     def tables(self):
         return [self.m_tables[tname] for tname in self.m_tables.keys()]
 
+    @staticmethod
+    def from_file(filename):
+        ys = CollectionYamlLoader(filename)
+        return ys.read_file()
+
+    @staticmethod
+    def from_dict(schema_dict):
+        ys = CollectionYamlLoader("dummy")
+        return ys._create_metadata_object(schema_dict)
+
+    def to_file(self, filename, collection_name):
+        ys = CollectionYamlLoader(filename)
+        ys.write_file(self, collection_name)
 
 """
     Common attributes for a table or a view
@@ -127,4 +141,144 @@ class Unknown:
     def typename(self):
         return "unknown"
 
+
+class CollectionYamlLoader:
+    def __init__(self, filename):
+        self.filename = filename
+
+    def read_file(self):
+        with open(self.filename, 'r') as stream:
+            try:
+                c_s = yaml.safe_load(stream)
+            except yaml.YAMLError as exc:
+                print(exc)
+        return self._create_metadata_object(c_s)
+
+    def _create_metadata_object(self, c_s):
+        keys = list(c_s.keys())
+        engine = "Unknown"
+
+        if len(keys) == 0:
+            raise ValueError("No collections in YAML file!")
+
+        if len(keys) == 1:
+            collection = keys[0]
+        elif len(keys) > 2:
+            raise ValueError("Please include only one collection per config file: " + str(keys))
+        else:  # we have two keys; one should be engine
+            if 'engine' not in keys:
+                raise ValueError("Please include only one collection per config file: " + str(keys))
+            engine = c_s['engine']
+            collection = [k for k in keys if k != 'engine'][0]
+
+        db = c_s[collection]
+
+        tables = []
+
+        for schema in db.keys():
+            s = db[schema]
+            for table in s.keys():
+                t = s[table]
+                tables.append(self.load_table(schema, table, t))
+
+        return CollectionMetadata(tables, engine)
+
+    def load_table(self, schema, table, t):
+        rowcount = int(t["rows"]) if "rows" in t else 0
+        row_privacy = bool(t["row_privacy"]) if "row_privacy" in t else False
+        max_ids = int(t["max_ids"]) if "max_ids" in t else 1
+        sample_max_ids = bool(t["sample_max_ids"]) if "sample_max_ids" in t else None
+        rows_exact = int(t["rows_exact"]) if "rows_exact" in t else None
+        clamp_counts = bool(t["clamp_counts"]) if "clamp_counts" in t else True
+
+        columns = []
+        colnames = [cn for cn in t.keys() if cn != "rows"]
+        for column in colnames:
+            columns.append(self.load_column(column, t[column]))
+
+        return Table(schema, table, rowcount, columns, row_privacy, max_ids, sample_max_ids, rows_exact, clamp_counts)
+
+    def load_column(self, column, c):
+        is_key = False if "private_id" not in c else bool(c["private_id"])
+        bounded = False if "bounded" not in c else bool(c["bounded"])
+
+        if c["type"] == "boolean":
+            return Boolean(column, is_key, bounded)
+        elif c["type"] == "datetime":
+            return DateTime(column, is_key, bounded)
+        elif c["type"] == "int":
+            minval = int(c["lower"]) if "lower" in c else None
+            maxval = int(c["upper"]) if "upper" in c else None
+            return Int(column, minval, maxval, is_key, bounded)
+        elif c["type"] == "float":
+            minval = float(c["lower"]) if "lower" in c else None
+            maxval = float(c["upper"]) if "upper" in c else None
+            return Float(column, minval, maxval, is_key, bounded)
+        elif c["type"] == "string":
+            card = int(c["cardinality"]) if "cardinality" in c else 0
+            return String(column, card, is_key, bounded)
+        else:
+            raise ValueError("Unknown column type for column {0}: {1}".format(column, c))
+
+    def write_file(self, collection_metadata, collection_name):
+
+        engine = collection_metadata.engine
+        schemas = {}
+        for t in collection_metadata.tables():
+            schema_name = t.schema
+            table_name = t.name
+            if schema_name not in schemas:
+                schemas[schema_name] = {}
+            schema = schemas[schema_name]
+            if table_name in schema:
+                raise ValueError("Attempt to insert table with same name twice: " + schema_name + table_name)
+            schema[table_name] = {}
+            table = schema[table_name]
+            table["rows"] = t.rowcount
+            if t.row_privacy is not None: 
+                table["row_privacy"] = t.row_privacy
+            if t.max_ids is not None: 
+                table["max_ids"] = t.max_ids
+            if t.sample_max_ids is not None: 
+                table["sample_max_ids"] = t.sample_max_ids
+            if t.rows_exact is not None:
+                table["rows_exact"] = t.rows_exact
+            if not t.clamp_counts:
+                table["clamp_counts"] = t.clamp_counts
+
+            for c in t.columns():
+                cname = c.name
+                if cname in table:
+                    raise ValueError("Duplicate column name {0} in table {1}".format(cname, table_name))
+                table[cname] = {}
+                column = table[cname]
+                if hasattr(c, "bounded") and c.bounded == True:
+                    column["bounded"] = c.bounded
+                if hasattr(c, "card"):
+                    column["cardinality"] = c.card
+                if hasattr(c, "minval") and c.minval is not None:
+                    column["lower"] = c.minval
+                if hasattr(c, "maxval") and c.maxval is not None:
+                    column["upper"] = c.maxval
+                if c.is_key is not None and c.is_key == True:
+                    column["private_id"] = c.is_key
+                if type(c) is String:
+                    column["type"] = "string"
+                elif type(c) is Int:
+                    column["type"] = "int"
+                elif type(c) is Float:
+                    column["type"] = "float"
+                elif type(c) is Boolean:
+                    column["type"] = "boolean"
+                elif type(c) is DateTime:
+                    column["type"] = "datetime"
+                elif type(c) is Unknown:
+                    column["type"] = "unknown"
+                else:
+                    raise ValueError("Unknown column type: " + str(type(c)))
+        db = {}
+        db[collection_name] = schemas
+        db["engine"] = collection_metadata.engine
+        with open(self.filename, 'w') as outfile:
+            yaml.dump(db, outfile)
