+++ sdk/burdock/query/sql/__init__.py
@@ -1,4 +1,4 @@
-from burdock.metadata.config_file import MetadataLoader
+from burdock.metadata.collection import CollectionMetadata
 from .ast.parse import QueryParser
 from .ast.validate import Validate
 from .private_rewrite import Rewriter
@@ -11,5 +11,5 @@ def execute_private_query(reader, schema, budget, query):
     query = reader._sanitize_query(query) if hasattr(reader ,"_sanitize_query") else query
     return PrivateQuery(reader, schema, budget).execute(query)
 
-__all__ = ["MetadataLoader", "QueryParser", "Rewriter", "Validate",
+__all__ = ["CollectionMetadata", "QueryParser", "Rewriter", "Validate",
            "PandasReader", "execute_private_query"]
