+++ tests/metadata/test_datasource_literal.py
@@ -12,6 +12,6 @@ class TestDSMetadataLiteral:
 
         table2 = copy.copy(table1)
         table2.name = "d2"
-        x = Collection([table1],"csv")
-        y = Collection([table2],"csv")
+        x = CollectionMetadata([table1],"csv")
+        y = CollectionMetadata([table2],"csv")
         assert(x["dbo.d1"].name != y["dbo.d2"].name)
